requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-docx==1.1.0
pdf2docx==0.5.8
pypandoc==1.12
Pillow==10.2.0
cairosvg==2.7.1
pillow-heif==0.15.0
reportlab==4.0.9
weasyprint==60.2
markdown==3.5.2
pdf2image==1.17.0
python-magic==0.4.27
werkzeug==3.0.1
PyMuPDF==1.23.26
weasyprint